requirements.txt
Django==2.0
djangorestframework==3.9.4
psycopg2==2.8.3
pytz==2019.1
uWSGI==2.0.18
